move physio file parameter queries
